pyproject.toml
[tool.poetry]
name = "userroutes"
version = "0.1.0"
description = ""
authors = ["kirill <k.otvetchikov@zebrains.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
structlog = "^23.1.0"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
alembic = "^1.12.0"
passlib = "^1.7.4"
jwt = "^1.3.1"
psycopg2-binary = "^2.9.7"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
